+++ CHANGELOG.md
@@ -1,3 +1,81 @@
+# (Mon Aug 10 2026)
+
+#### 🐛 Bug Fix
+
+- `@quenty/remoting@12.41.1`, `@quenty/rogue-properties@11.57.1`
+  - Users/euvinkeel/memleak2 [#789](https://github.com/Quenty/NevermoreEngine/pull/789) ([@euvinkeel](https://github.com/euvinkeel))
+
+#### Authors: 1
+
+- euvinkeel ([@euvinkeel](https://github.com/euvinkeel))
+
+---
+
+# (Sat Aug 08 2026)
+
+#### 🐛 Bug Fix
+
+- `@quenty/gameproductservice@14.66.9`
+  - fix: Server pass purchase validation gate [#792](https://github.com/Quenty/NevermoreEngine/pull/792) ([@alex-y-z](https://github.com/alex-y-z))
+
+#### Authors: 1
+
+- Alex Turner ([@alex-y-z](https://github.com/alex-y-z))
+
+---
+
+# (Sat Aug 08 2026)
+
+#### 🐛 Bug Fix
+
+- `@quenty/gameproductservice@14.66.8`
+  - fix: Validate game pass ownership on server [#791](https://github.com/Quenty/NevermoreEngine/pull/791) ([@alex-y-z](https://github.com/alex-y-z))
+
+#### Authors: 1
+
+- Alex Turner ([@alex-y-z](https://github.com/alex-y-z))
+
+---
+
+# (Fri Aug 07 2026)
+
+#### 🐛 Bug Fix
+
+- `@quenty/highlight@10.49.1`, `@quenty/performanceutils@4.12.1`, `@quenty/sounds@10.27.1`
+  - Update PerformanceUtils.lua [#787](https://github.com/Quenty/NevermoreEngine/pull/787) ([@euvinkeel](https://github.com/euvinkeel))
+
+#### Authors: 1
+
+- euvinkeel ([@euvinkeel](https://github.com/euvinkeel))
+
+---
+
+# (Fri Aug 07 2026)
+
+#### 🐛 Bug Fix
+
+- `@quenty/clienttranslator@14.47.5`
+  - fix: Prevent more redundant localization table rebuilding [#788](https://github.com/Quenty/NevermoreEngine/pull/788) ([@alex-y-z](https://github.com/alex-y-z))
+
+#### Authors: 1
+
+- Alex Turner ([@alex-y-z](https://github.com/alex-y-z))
+
+---
+
+# (Wed Aug 05 2026)
+
+#### 🐛 Bug Fix
+
+- `@quenty/clienttranslator@14.47.4`
+  - fix: Prevent redundant localization table rebuilding [#786](https://github.com/Quenty/NevermoreEngine/pull/786) ([@alex-y-z](https://github.com/alex-y-z))
+
+#### Authors: 1
+
+- Alex Turner ([@alex-y-z](https://github.com/alex-y-z))
+
+---
+
 # (Thu Jul 30 2026)
 
 #### 🐛 Bug Fix