+++ src/access/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.12](https://github.com/Quenty/NevermoreEngine/compare/@quenty/access@1.3.11...@quenty/access@1.3.12) (2026-08-10)
+
+**Note:** Version bump only for package @quenty/access
+
+## [1.3.11](https://github.com/Quenty/NevermoreEngine/compare/@quenty/access@1.3.10...@quenty/access@1.3.11) (2026-08-08)
+
+**Note:** Version bump only for package @quenty/access
+
+## [1.3.10](https://github.com/Quenty/NevermoreEngine/compare/@quenty/access@1.3.9...@quenty/access@1.3.10) (2026-08-08)
+
+**Note:** Version bump only for package @quenty/access
+
+## [1.3.9](https://github.com/Quenty/NevermoreEngine/compare/@quenty/access@1.3.8...@quenty/access@1.3.9) (2026-08-07)
+
+**Note:** Version bump only for package @quenty/access
+
+## [1.3.8](https://github.com/Quenty/NevermoreEngine/compare/@quenty/access@1.3.7...@quenty/access@1.3.8) (2026-08-05)
+
+**Note:** Version bump only for package @quenty/access
+
 ## [1.3.7](https://github.com/Quenty/NevermoreEngine/compare/@quenty/access@1.3.6...@quenty/access@1.3.7) (2026-07-30)
 
 **Note:** Version bump only for package @quenty/access