+++ src/animationprovider/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [11.48.1](https://github.com/Quenty/NevermoreEngine/compare/@quenty/animationprovider@11.48.0...@quenty/animationprovider@11.48.1) (2026-08-10)
+
+**Note:** Version bump only for package @quenty/animationprovider
+
 # [11.48.0](https://github.com/Quenty/NevermoreEngine/compare/@quenty/animationprovider@11.47.2...@quenty/animationprovider@11.48.0) (2026-07-28)
 
 **Note:** Version bump only for package @quenty/animationprovider