+++ src/humanoidspeed/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [12.61.1](https://github.com/Quenty/NevermoreEngine/compare/@quenty/humanoidspeed@12.61.0...@quenty/humanoidspeed@12.61.1) (2026-08-10)
+
+**Note:** Version bump only for package @quenty/humanoidspeed
+
 # [12.61.0](https://github.com/Quenty/NevermoreEngine/compare/@quenty/humanoidspeed@12.60.2...@quenty/humanoidspeed@12.61.0) (2026-07-28)
 
 **Note:** Version bump only for package @quenty/humanoidspeed