+++ src/scoredactionservice/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [16.54.5](https://github.com/Quenty/NevermoreEngine/compare/@quenty/scoredactionservice@16.54.4...@quenty/scoredactionservice@16.54.5) (2026-08-07)
+
+**Note:** Version bump only for package @quenty/scoredactionservice
+
+## [16.54.4](https://github.com/Quenty/NevermoreEngine/compare/@quenty/scoredactionservice@16.54.3...@quenty/scoredactionservice@16.54.4) (2026-08-05)
+
+**Note:** Version bump only for package @quenty/scoredactionservice
+
 ## [16.54.3](https://github.com/Quenty/NevermoreEngine/compare/@quenty/scoredactionservice@16.54.2...@quenty/scoredactionservice@16.54.3) (2026-07-29)
 
 **Note:** Version bump only for package @quenty/scoredactionservice