+++ src/secrets/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.64.2](https://github.com/Quenty/NevermoreEngine/compare/@quenty/secrets@7.64.1...@quenty/secrets@7.64.2) (2026-08-10)
+
+**Note:** Version bump only for package @quenty/secrets
+
 ## [7.64.1](https://github.com/Quenty/NevermoreEngine/compare/@quenty/secrets@7.64.0...@quenty/secrets@7.64.1) (2026-07-30)
 
 **Note:** Version bump only for package @quenty/secrets