+++ src/settings-inputkeymap/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [10.70.7](https://github.com/Quenty/NevermoreEngine/compare/@quenty/settings-inputkeymap@10.70.6...@quenty/settings-inputkeymap@10.70.7) (2026-08-10)
+
+**Note:** Version bump only for package @quenty/settings-inputkeymap
+
+## [10.70.6](https://github.com/Quenty/NevermoreEngine/compare/@quenty/settings-inputkeymap@10.70.5...@quenty/settings-inputkeymap@10.70.6) (2026-08-07)
+
+**Note:** Version bump only for package @quenty/settings-inputkeymap
+
+## [10.70.5](https://github.com/Quenty/NevermoreEngine/compare/@quenty/settings-inputkeymap@10.70.4...@quenty/settings-inputkeymap@10.70.5) (2026-08-05)
+
+**Note:** Version bump only for package @quenty/settings-inputkeymap
+
 ## [10.70.4](https://github.com/Quenty/NevermoreEngine/compare/@quenty/settings-inputkeymap@10.70.3...@quenty/settings-inputkeymap@10.70.4) (2026-07-30)
 
 **Note:** Version bump only for package @quenty/settings-inputkeymap