+++ src/settings/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [11.67.2](https://github.com/Quenty/NevermoreEngine/compare/@quenty/settings@11.67.1...@quenty/settings@11.67.2) (2026-08-10)
+
+**Note:** Version bump only for package @quenty/settings
+
 ## [11.67.1](https://github.com/Quenty/NevermoreEngine/compare/@quenty/settings@11.67.0...@quenty/settings@11.67.1) (2026-07-30)
 
 **Note:** Version bump only for package @quenty/settings