+++ src/softshutdown/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [10.3.8](https://github.com/Quenty/NevermoreEngine/compare/@quenty/softshutdown@10.3.7...@quenty/softshutdown@10.3.8) (2026-08-10)
+
+**Note:** Version bump only for package @quenty/softshutdown
+
+## [10.3.7](https://github.com/Quenty/NevermoreEngine/compare/@quenty/softshutdown@10.3.6...@quenty/softshutdown@10.3.7) (2026-08-07)
+
+**Note:** Version bump only for package @quenty/softshutdown
+
+## [10.3.6](https://github.com/Quenty/NevermoreEngine/compare/@quenty/softshutdown@10.3.5...@quenty/softshutdown@10.3.6) (2026-08-05)
+
+**Note:** Version bump only for package @quenty/softshutdown
+
 ## [10.3.5](https://github.com/Quenty/NevermoreEngine/compare/@quenty/softshutdown@10.3.4...@quenty/softshutdown@10.3.5) (2026-07-30)
 
 **Note:** Version bump only for package @quenty/softshutdown